+++ ast/ast.go
@@ -3,6 +3,7 @@ package ast
 import (
 	"context"
 	"fmt"
+	"io"
 	"strings"
 
 	"github.com/antlr/antlr4/runtime/Go/antlr"
@@ -14,7 +15,38 @@ import (
 
 // Parse parses an earthfile into an AST.
 func Parse(ctx context.Context, filePath string, enableSourceMap bool) (ef spec.Earthfile, err error) {
-	version, err := ParseVersion(filePath, enableSourceMap)
+	var opts []Opt
+	if enableSourceMap {
+		opts = append(opts, WithSourceMap())
+	}
+	return ParseOpts(ctx, FromPath(filePath), opts...)
+}
+
+// ParseOpts parses an earthfile into an AST. This is the functional option
+// version, which uses option functions to change how a file is parsed.
+func ParseOpts(ctx context.Context, from FromOpt, opts ...Opt) (spec.Earthfile, error) {
+	defaultPrefs := prefs{
+		done: func() {},
+	}
+	prefs, err := from(defaultPrefs)
+	if err != nil {
+		return spec.Earthfile{}, errors.Wrap(err, "ast: could not apply FromOpt")
+	}
+	for _, opt := range opts {
+		newPrefs, err := opt(prefs)
+		if err != nil {
+			return spec.Earthfile{}, errors.Wrap(err, "ast: could not apply options")
+		}
+		prefs = newPrefs
+	}
+
+	defer prefs.done()
+
+	var versionOpts []Opt
+	if prefs.enableSourceMap {
+		versionOpts = append(versionOpts, WithSourceMap())
+	}
+	version, err := ParseVersionOpts(FromReader(prefs.reader), versionOpts...)
 	if err != nil {
 		return spec.Earthfile{}, err
 	}
@@ -22,13 +54,21 @@ func Parse(ctx context.Context, filePath string, enableSourceMap bool) (ef spec.
 	// Convert.
 	errorListener := antlrhandler.NewReturnErrorListener()
 	errorStrategy := antlrhandler.NewReturnErrorStrategy()
-	tree, err := newEarthfileTree(filePath, errorListener, errorStrategy)
+
+	if _, err := prefs.reader.Seek(0, 0); err != nil {
+		return spec.Earthfile{}, errors.Wrap(err, "ast: could not seek to beginning of file")
+	}
+	b, err := io.ReadAll(prefs.reader)
+	if err != nil {
+		return spec.Earthfile{}, errors.Wrap(err, "ast: could not read Earthfile for parsing")
+	}
+	stream, tree, err := newEarthfileTree(string(b), errorListener, errorStrategy)
 	if err != nil {
 		return spec.Earthfile{}, err
 	}
-	ef, walkErr := walkTree(newListener(ctx, filePath, enableSourceMap), tree)
+	ef, walkErr := walkTree(newListener(ctx, stream, prefs.reader.Name(), prefs.enableSourceMap), tree)
 	if len(errorListener.Errs) > 0 {
-		errString := []string{fmt.Sprintf("lexer error: %s", filePath)}
+		errString := []string{fmt.Sprintf("lexer error: %s", prefs.reader.Name())}
 		for _, err := range errorListener.Errs {
 			errString = append(errString, err.Error())
 		}
@@ -37,7 +77,7 @@ func Parse(ctx context.Context, filePath string, enableSourceMap bool) (ef spec.
 	if errorStrategy.Err != nil {
 		return spec.Earthfile{}, errors.Wrapf(
 			errorStrategy.Err, "%s line %d:%d '%s'",
-			filePath,
+			prefs.reader.Name(),
 			errorStrategy.RE.GetOffendingToken().GetLine(),
 			errorStrategy.RE.GetOffendingToken().GetColumn(),
 			errorStrategy.RE.GetOffendingToken().GetText())
@@ -57,28 +97,24 @@ func Parse(ctx context.Context, filePath string, enableSourceMap bool) (ef spec.
 
 func walkTree(l *listener, tree parser.IEarthFileContext) (spec.Earthfile, error) {
 	antlr.ParseTreeWalkerDefault.Walk(l, tree)
-	err := l.Err()
-	if err != nil {
+	if err := l.Err(); err != nil {
 		return spec.Earthfile{}, err
 	}
 	return l.Earthfile(), nil
 }
 
-func newEarthfileTree(filename string, errorListener *antlrhandler.ReturnErrorListener, errorStrategy antlr.ErrorStrategy) (parser.IEarthFileContext, error) {
-	input, err := antlr.NewFileStream(filename)
-	if err != nil {
-		return nil, errors.Wrapf(err, "new file stream %s", filename)
-	}
+func newEarthfileTree(body string, errorListener *antlrhandler.ReturnErrorListener, errorStrategy antlr.ErrorStrategy) (*antlr.CommonTokenStream, parser.IEarthFileContext, error) {
+	input := antlr.NewInputStream(body)
 	lexer := newLexer(input)
 	lexer.RemoveErrorListeners()
 	lexer.AddErrorListener(errorListener)
 	stream := antlr.NewCommonTokenStream(lexer, 0)
 	if lexer.Err() != nil {
-		return nil, lexer.Err()
+		return nil, nil, lexer.Err()
 	}
 	p := parser.NewEarthParser(stream)
 	p.AddErrorListener(errorListener)
 	p.SetErrorHandler(errorStrategy)
 	p.BuildParseTrees = true
-	return p.EarthFile(), nil
+	return stream, p.EarthFile(), nil
 }