+++ cmd/earthly/root_cmds.go
@@ -16,6 +16,7 @@ import (
 	"github.com/urfave/cli/v2"
 	"golang.org/x/sync/errgroup"
 
+	"github.com/earthly/earthly/ast/spec"
 	"github.com/earthly/earthly/buildcontext"
 	"github.com/earthly/earthly/buildkitd"
 	"github.com/earthly/earthly/config"
@@ -24,6 +25,7 @@ import (
 	"github.com/earthly/earthly/earthfile2llb"
 	"github.com/earthly/earthly/util/cliutil"
 	"github.com/earthly/earthly/util/fileutil"
+	"github.com/earthly/earthly/util/platutil"
 	"github.com/earthly/earthly/util/termutil"
 )
 
@@ -114,8 +116,14 @@ func (app *earthlyApp) rootCmds() []*cli.Command {
 			Subcommands: app.orgCmds(),
 		},
 		{
+			Name:      "doc",
+			Usage:     "Document targets from an Earthfile",
+			UsageText: "earthly [options] doc [<project-ref>[+<target-ref>]]",
+			Action:    app.actionDocumentTarget,
+		},
+		{
 			Name:      "ls",
-			Usage:     "List targets from an Earthfile *experimental*",
+			Usage:     "List targets from an Earthfile",
 			UsageText: "earthly [options] ls [<project-ref>]",
 			Action:    app.actionListTargets,
 			Flags: []cli.Flag{
@@ -561,6 +569,87 @@ func ifNilBoolDefault(ptr *bool, defaultValue bool) bool {
 	return *ptr
 }
 
+func (app *earthlyApp) actionDocumentTarget(cliCtx *cli.Context) error {
+	app.commandName = "docTarget"
+
+	if cliCtx.NArg() > 1 {
+		return errors.New("invalid number of arguments provided")
+	}
+
+	var tgtPath string
+	if cliCtx.NArg() > 0 {
+		tgtPath = cliCtx.Args().Get(0)
+		switch tgtPath[0] {
+		case '.', '/', '+':
+		default:
+			return errors.New("remote-paths are not currently supported - documentation targets must start with one of ['.', '/', '+']")
+		}
+	}
+
+	singleTgt := true
+	if !strings.ContainsRune(tgtPath, '+') {
+		tgtPath += "+base"
+		singleTgt = false
+	}
+
+	target, err := domain.ParseTarget(tgtPath)
+	if err != nil {
+		return errors.Errorf("unable to parse target [%v]", tgtPath)
+	}
+
+	gitLookup := buildcontext.NewGitLookup(app.console, app.sshAuthSock)
+	resolver := buildcontext.NewResolver(nil, gitLookup, app.console, "")
+	platr := platutil.NewResolver(platutil.GetUserPlatform())
+	var gwClient gwclient.Client
+	bc, err := resolver.Resolve(cliCtx.Context, gwClient, platr, target)
+	if err != nil {
+		return errors.Wrap(err, "failed to resolve target")
+	}
+
+	const docsIndent = "  "
+
+	if singleTgt {
+		tgt, err := findTarget(bc.Earthfile, target.Target)
+		if err != nil {
+			return errors.Wrap(err, "failed to look up target")
+		}
+		return app.documentSingleTarget("", docsIndent, tgt)
+		// TODO: print arg docs - probably a flag to documentSingleTarget.
+	}
+
+	tgts := bc.Earthfile.Targets
+	fmt.Println("TARGETS:")
+	const (
+		nameIndent = docsIndent
+		bodyIndent = docsIndent + docsIndent
+	)
+	for _, tgt := range tgts {
+		_ = app.documentSingleTarget(nameIndent, bodyIndent, tgt)
+	}
+
+	return nil
+}
+
+func (app *earthlyApp) documentSingleTarget(nameIndent, docIndent string, tgt spec.Target) error {
+	if tgt.Docs == "" {
+		return errors.Errorf("no doc comment found [hint: add a comment starting with the word '%s' on the line immediately above this target]", tgt.Name)
+	}
+
+	firstWordEnd := strings.IndexRune(tgt.Docs, ' ')
+	if firstWordEnd == -1 {
+		return errors.Errorf("failed to parse first word of documentation comments")
+	}
+	firstWord := tgt.Docs[:firstWordEnd]
+	if firstWord != tgt.Name {
+		return errors.Errorf("no doc comment found [hint: a comment was found but the first word was not '%s']", tgt.Name)
+	}
+
+	fmt.Println(indent(nameIndent, "+"+tgt.Name))
+	indented := indent(docIndent, tgt.Docs)
+	fmt.Println(strings.Trim(indented, "\n"))
+	return nil
+}
+
 func (app *earthlyApp) actionListTargets(cliCtx *cli.Context) error {
 	app.commandName = "listTargets"
 
@@ -697,4 +786,24 @@ func (app *earthlyApp) actionPrune(cliCtx *cli.Context) error {
 	}
 	app.console.Printf("Freed %s\n", humanize.Bytes(total))
 	return nil
+}
+
+func findTarget(ef spec.Earthfile, name string) (spec.Target, error) {
+	for _, tgt := range ef.Targets {
+		if tgt.Name == name {
+			return tgt, nil
+		}
+	}
+	return spec.Target{}, errors.Errorf("could not find target named [%v]", name)
+}
+
+func indent(indent, s string) string {
+	lines := strings.Split(s, "\n")
+	for i, l := range lines {
+		if l == "" {
+			continue
+		}
+		lines[i] = indent + l
+	}
+	return strings.Join(lines, "\n")
 }